% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_criterion}
\alias{calculate_criterion}
\title{Calculate five clear sky criterion.}
\arguments{
\item{x}{Numeric vector of measured irradiance.}

\item{cs}{Numeric vector of predicted irradiance. Should be the same
length as x, however this condition is not checked.}
}
\description{
Calculate five clear sky criterion.
}
\keyword{internal}

