% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clearskymethods.R
\name{clear_points}
\alias{clear_points}
\alias{clear_points.clearsky}
\alias{clear_points.default}
\title{Clear sky detection}
\usage{
clear_points(x, ...)

\method{clear_points}{default}(x, cs, thresholds, window_len, ...)

\method{clear_points}{clearsky}(x, thresholds, window_len, ...)
}
\arguments{
\item{x}{Numeric vector of measured irradiance values or object of clear_sky
class containing both observed and predicted members.}

\item{...}{ignored.}

\item{cs}{Numeric vector of predicted irradiance from a clear sky model.}

\item{thresholds}{List of vectors, each vector containing the threshold
values for their respective clear sky criteria. Each vector must have length
greater than or equal to two, with the minimum and maximum values in the
vector being used as thresholds. The list must be arranged in the following
order:
\enumerate{
    \item Mean
    \item Max
    \item Line length
    \item Sigma
    \item Maximum deviation from clear sky slope
}}

\item{window_len}{Length of window to use in calculating criterion. Must be
a positive integer.}
}
\value{
The form of the value returned by 'clear_points' depends on the class
of its argument.

The default method returns a logical vector of the same length as x. TRUE
indicates that the corresponding measured irradiance value is clear.

The clearsky method returns a clearsky object with member 'clear' set to a
logical vector of the same length as x.
}
\description{
Implementation of the clear sky detection algorithm found in Reno et al,
2012. (See References). Determines clear points using a rolling window and
five criterion.
}
\details{
A point is declared clear if it is determined to be clear at least once.
}
\section{References}{

Global Horizontal Irradiance Clear Sky Models: Implementation and Analysis,
Reno et al, 2012, pp. 28-36.
}

