% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clear_pts}
\alias{clear_pts}
\title{Clear sky detection.}
\usage{
clear_pts(x, cs, thresholds, window_len)
}
\arguments{
\item{x}{Numeric vector of measured irradiance values.}

\item{cs}{Numeric vector of predicted irradiance from a clear sky model.}

\item{thresholds}{List of vectors, each vector containing the threshold
values for their respective clear sky criteria. Each vector must have length
greater than or equal to two, with the minimum and maximum values in the
vector being used as thresholds. The list must be arranged in the following
order:
\enumerate{
    \item Mean
    \item Max
    \item Line length
    \item Sigma
    \item Maximum deviation from clear sky slope
}}

\item{window_len}{Length of window, in minutes, used in calculating
criterion. Must be a positive integer.}
}
\value{
A logical vector of the same length as x, TRUE indicates the
point is clear.

A logical vector of the same length as x. TRUE indicates that the
corresponding measured irradiance value in x is clear.
}
\description{
Determine clear points using a rolling window and five clear sky criterion.
}
\details{
A point is declared clear if it is determined to be clear at least once.
}
\section{References}{

Global Horizontal Irradiance Clear Sky Models: Implementation and Analysis,
Reno et al, 2012, pp. 28-36.
}

