% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clearskymodels.R
\name{clear_sky}
\alias{clear_sky}
\title{Fit clear sky model.}
\usage{
clear_sky(model, x, y, ...)
}
\arguments{
\item{model}{Name of model to be fit.}

\item{x}{Dataframe, named list or named vector containing date information
for model. Must have elements named 'DayOfYear', 'Year' and 'Interval'.}

\item{y}{Dataframe, named list or named vector containing location information
for model. Must have elements named 'Timezone' or 'TZ', 'Longitude',
'Latitude' and, if necessary for the model, 'Elevation'.}

\item{...}{Additional arguments to the model. Normally a named vector of
model parameters.}
}
\value{
A vector of containing the modeled GHI values for each day at each
specified interval.
}
\description{
Fit clear sky model.
}
\examples{
# Fit Robledo-Soler model to the year of 2014 for Eugene, Oregon
time_info <- data.frame(DayOfYear = 1:365, Year = rep(2014, 365),
                        Interval = rep(1, 365))
location_info <- data.frame(Latitude = 44.05, Longitude = -123.07,
                            Elevation = 150, TZ = -8)
model_params <- c(a = 1000, b = 1.5, c = 0.2)

model <- clear_sky('RS', x = time_info, y = location_info,
                   parameters = model_params)
}

