% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkginfo.R
\docType{package}
\name{clearskies}
\alias{clearskies}
\alias{clearskies-package}
\title{clearskies: R package for fitting clear sky models and detecting clear
points.}
\description{
The clearskies package contains functions for fitting a number of clear sky
models and a function for running the clear sky detection algorithm.
}
\section{Clear Sky Models}{

Clear sky models are fit through the clear_sky function, which provides a
convenient interface to all available clear sky models. Models can be fit by
passing either a dataframe, named list or named vector of date information to
x and a named object of location information to y, or by explicitly passing
the necessary date and location parameters to their respective arguments.
Parameters specific to the model type may be passed through a named object to
argument "parameters".
}

\section{Clear Sky Detection}{

The clear sky detection algorithm can be run on measured irradiance and a
corresponding clear sky model using the clear_points function. clear_points
requires that the thresholds argument is ordered correctly, see ?clear_points
for details. Thresholds for 10-minute rolling windows can be found in data/.

More information on the clear sky detection algorithm can found in:
Global Horizontal Irradiance Clear Sky Models: Implementation and Analysis,
Reno et al, 2012, pp. 28.
}

