% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkginfo.R
\docType{data}
\name{eugene}
\alias{eugene}
\title{Sample solar irradiance dataset.}
\format{A data.frame with 7 columns:
\describe{
     \item{Year}{2014}
     \item{Month}{September}
     \item{DayOfMonth}{The day of the month, 1 through 30}
     \item{DayOfYear}{The day of the year, 244 through 273}
     \item{Interval}{Time between observations in minutes}
     \item{Minutes}{Minutes elapsed since midnight}
     \item{GHI}{Measured global horizontal solar irradiance}
}}
\source{
Raw data available through the University of Oregon Solar Radiation
Monitoring Lab website:

http://solardat.uoregon.edu/SelectArchival.html
}
\usage{
eugene
}
\description{
A dataset of measured global horizontal irradiance (GHI) from Eugene, Oregon
for the month of September, 2014.
}
\keyword{datasets}

