% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{evaluate_criterion}
\alias{evaluate_criterion}
\title{Check if all criterion are within their respective threshold values.}
\arguments{
\item{criterion}{List of clear sky criterion.}

\item{thresholds}{List of vectors, each vector containing the threshold
values for their respective clear sky criteria.}
}
\value{
If all criterion are within their threshold ranges, inclusive,
returns true.

Note: Clear sky criterion from \emph{criterion} are compared to values in
\emph{thresholds} by index, not name.
}
\description{
Check if all criterion are within their respective threshold values.
}
\keyword{internal}

