% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exrad}
\alias{exrad}
\title{Helper function for calculating solar irradiance in clear sky models.}
\usage{
exrad(dayofyear, times)
}
\arguments{
\item{dayofyear}{Day of year to calculate solar irradiance for. May be a
vector.}

\item{times}{Length of return vector. Passed to \emph{rep}.}
}
\value{
A vector of earth radius values of length times.
}
\description{
Helper function for calculating solar irradiance in clear sky models.
}
\keyword{internal}

