% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkginfo.R
\docType{data}
\name{locations}
\alias{locations}
\title{Location information for 10 sites in the Pacific Northwest.}
\format{A data frame with 10 rows and 5 variables:
\describe{
     \item{Site}{Location abbreviation}
     \item{Latitude}{Latitude coordinate}
     \item{Longitude}{Longitude coordinate}
     \item{Elevation}{Elevation at the location, in meters}
     \item{TZ}{UTC Offset}
}}
\usage{
locations
}
\description{
A dataset containing GPS coordinates, elevation and UTC Offset information
for 10 sites in the Pacific Northwest.
}
\keyword{datasets}

