% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{Numeric vector.}

\item{y}{Numeric vector.}
}
\value{
Square root of the mean squared error between x and y.
}
\description{
Calculate root mean squared error.
}

