% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{S}
\alias{S}
\title{Maximum deviation of the measured irradiance from the clear sky slope.}
\arguments{
\item{x}{Numeric vector of measured irradiance.}

\item{cs}{Numeric vector of predicted irradiance. Should be the same
length as x, however this condition is not checked.}
}
\description{
The maximum deviation of the measured irradiance from the clear sky slope is
one of five criterion for detecting clear points. It is defined as follows:
\deqn{
S = max\{|s_{i}-d_{i}|\}
}
where \eqn{ s_{i} = GHI_{i+1} - GHI_{i} } and \eqn{ d_{i} = y_{i+1} - y_{i}
}. \eqn{GHI} denotes the measured irradiance and \eqn{y} denotes the
predicted irradiance from a clear sky model.
}
\section{References}{

Global Horizontal Irradiance Clear Sky Models: Implementation and Analysis,
Reno et al, 2012, pp. 31.
}
\keyword{internal}

