% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigma}
\alias{sigma}
\title{Normalized standard deviation of the slope between sequential points.}
\arguments{
\item{x}{Numeric vector of measured irradiance.}
}
\description{
The normalized standard deviation of the slope (sigma) is one of the five
criterion for detecting clear points. It is defined as follows:
\deqn{
\sigma = \frac{1}{ {\overline{GHI}} }
\sqrt{ \frac{1}{n-1} \sum_{i=1}^{n-1} (s_{i} {\overline{s}})^2 }
}
where \eqn{ {\overline{GHI}} } is the mean of the measured irradiance
values, \eqn{ s_{i} = GHI_{i+1} - GHI_{i} } and \eqn{ {\overline{s}} } is the
sample standard deviation of \eqn{s}.
}
\section{References}{

Global Horizontal Irradiance Clear Sky Models: Implementation and Analysis,
Reno et al, 2012, pp. 30.
}
\keyword{internal}

