% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkginfo.R
\docType{data}
\name{thresholds}
\alias{thresholds}
\title{Clear sky detection thresholds for 10-minute windows.}
\format{A named list with 5 components:
\describe{
     \item{Mean}{Vector of minimum and maximum mean values}
     \item{Max}{Vector of minimum and maximum max values}
     \item{Line.length}{Vector of minimum and maximum line length values}
     \item{Sigma}{Vector of minimum and maximum sigma values}
     \item{Deviation}{Vector of minimum and maximum slope deviation values}
}}
\usage{
thresholds
}
\description{
Contains threshold values used for detecting clear points with a 10-minute
window.
}
\section{References}{

Global Horizontal Irradiance Clear Sky Models: Implementation and Analysis,
Reno et al, 2012, pp. 34.
}
\keyword{datasets}

