% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zenith}
\alias{zenith}
\title{Calculate the zenith angle.}
\usage{
zenith(dayofyear, year, tz, latitude, longitude, interval = 1L)
}
\arguments{
\item{dayofyear}{Numeric vector containing the day of year(s),
for which the zenith angle should be calculated.}

\item{year}{Numeric vector containing the year(s) for which the zenith
angle sould be calculated.}

\item{tz}{UTC Offset. Ex: Eastern Standard Time = -5.}

\item{latitude}{Latitude of the location at which the zenith angle is to
be calculated.}

\item{longitude}{Longitude of the location at which the zenith angle is to
be calculated.}

\item{interval}{Number of minutes between zenith angle calculations. Defaults to
(every minute). Must be an integer between 1 and 60, inclusive.}
}
\value{
A single vector of the zenith angles at each interval throughout the
specified time period
(i.e. a vector of length (60 * 24 / interval) * number of days).
}
\description{
Calculate the zenith angle.
}
\details{
zenith is vectorized over both dayofyear and year, with the shorter vector
being recycled as usual.
}
\keyword{internal}

